Added Worker Adding Pages

+++ ETrade/Entities/Concrete/ShippingDetail.cs
@@ -11,7 +11,7 @@ namespace Entities.Concrete
         public string AddressName { get; set; }
         public int CustomerId { get; set; }
         public byte CityId { get; set; }
-        public byte DistrictId { get; set; }
+        public short DistrictId { get; set; }
         public string Street { get; set; }
         public string Number { get; set; }
